fix: repair ops console page layout
